Updated with remaining features.

+++ Account/src/main/java/com/mavericsystems/demo/repository/AccountRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface AccountRepo extends CrudRepository <Account, Integer> {
+public interface AccountRepo extends CrudRepository <Account, Long> {
     List<Account> findByCustomerID(Integer customerID);
 }
